Adjust routing
